+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ## 0.2.3.0 (not released yet, TBD)
+- Minor improves in generated `.headroom.config` in `init` mode
 
 ## 0.2.2.0 (released 2020-05-04)
 - [#45] Add `-c|--check-headers` command line option